apps/api/requirements.txt
# FastAPI and ASGI
fastapi[all]==0.110.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Task Queue
celery[redis]==5.3.6
redis==5.0.1

# Auth & Security
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AWS & Storage
boto3==1.34.34
botocore==1.34.34

# Video Processing
ffmpeg-python==0.2.0

# AI & ML
anthropic==0.12.0
deepgram-sdk==3.2.7
mediapipe==0.10.9

# Third-party APIs
stripe==7.12.0
requests==2.31.0

# Utilities & Config
pydantic==2.5.3
pydantic-settings==2.1.0
structlog==23.2.0
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-xdist==3.5.0
httpx==0.26.0

# Windows compatibility
pywin32==306; sys_platform == "win32"

# Missing dependencies for JWT/Auth
pyjwt==2.8.0
cryptography==42.0.2

# Testing with SQLite for tests
aiosqlite==0.19.0